Pre-Launch Shriram Plots Sarjapur Bangalore
Shriram Plots Sarjapur is one of the prime developments in Bangalore that offers plots of diverse extents. The complex has a huge open space inside the project and will feature a grand clubhouse with all the cutting-edge conveniences.

The project is going to offer a stream of luxurious amenities including a swimming pool where residents can relax and refresh. There is a practice room, yoga room, badminton court, indoor and outdoor games area, a kid’s pool, an adolescent play area, 24-hour water supply, 24-hour security, and so on.

The Sarjapur Road offers a number of job opportunities which has led to an increase in residential properties around the range. Shriram Sarjapur Plots is connected with different parts of Bengaluru. A number of leading Schools, Hospitals, Bus stops, Malls, and shopping complexes are located close to the development.

Developed by Shriram Properties one of the leading developers in Bangalore the project is filled with all modern social and civic conveniences. Shriram Properties is one of the most reputed and has a presence in numerous towns. The brand has a legacy of quality and belief. Home buyers looking to invest in plots in the electronic city must consider the development. The locality, as well as the range of facilities available, make it one of the prime investment options for purchasers.

Project Advantages:-

✔ Gated community with amenity space for the residents
✔ A Centrally-located Retail Plaza & Town Square
✔ 2 Luxurious Grand Clubhouses	
✔ Plots ranging from 100 sq. yd to 180 sq. yd

Panorama of Breathtaking Splendor Amenities:-

✔ Swimming Pool	
✔ Kids Play Area
✔ Gym	
✔ Walking Path
✔ Convenience Store	
✔ Multipurpose Hall
✔ Spa	
✔ Reflexology Track

For More Details:
Visit Here: https://bit.ly/3RPHbPl